refactor: Settings 파일 구조 조정
