+++ src/app/routes/HOME_ROUTES.tsx
@@ -11,7 +11,7 @@ import { RouterDOM } from "@app/types/common.ts";
 import DetailSetting from "../../legacies/assetManagement/SavingsGoalContainer/DetailSetting";
 import FetchPaymentHistory from "../../pages/Home/components/HomeContainer/view/FetchPaymentHistory";
 import SearchSchedule from "../../pages/Home/components/HomeContainer/view/SearchSchedule";
-import MyData from "@pages/Settings/connection/MyData";
+import MyData from "@pages/Settings/components/connection/MyData";
 import ScheduleListPage from "@pages/Home/pages/ScheduleListPage/ScheduleListPage.tsx";
 
 const HOME_ROUTES: RouterDOM[] = [
