--- src/components/settings/ClickableListItem/index.tsx
@@ -1,3 +1,0 @@
-import ClickableListItem from './ClickableListItem';
-
-export default ClickableListItem;
