--- src/components/settings/ToggleListItem/index.tsx
@@ -1,3 +1,0 @@
-import ToggleListItem from './ToggleListItem';
-
-export default ToggleListItem;
