+++ src/pages/Settings/components/ClickableListItem/index.tsx
@@ -1,0 +1,3 @@
+import ClickableListItem from "./ClickableListItem.tsx";
+
+export default ClickableListItem;
