+++ src/pages/Settings/components/ToggleListItem/index.tsx
@@ -1,0 +1,3 @@
+import ToggleListItem from "./ToggleListItem.tsx";
+
+export default ToggleListItem;
