+++ src/pages/Settings/components/display/ThemeMode.tsx
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import ToggleListItem from "@components/settings/ToggleListItem";
+import ToggleListItem from "pages/Settings/components/ToggleListItem";
 import {
   changeThemeMode,
   selectIsDarkMode,
