+++ src/pages/Settings/components/version/Version/index.tsx
